feat: add time traveller
