+++ test/rskRegistrar.test.ts
@@ -2,6 +2,7 @@ import { RSKRegistrar } from '../src/RSKRegistrar'
 
 // @ts-ignore
 import { deployRskRegistrar } from './util'
+import { providers } from 'ethers'
 
 describe('rsk registrar', () => {
   test('dummy', async () => {
@@ -21,17 +22,28 @@ describe('RSKRegistrar SDK', () => {
     console.log('fifsAddrRegistrar.address: ', fifsAddrRegistrar.address)
     console.log('rifToken.address: ', rifToken.address)
     const salt = generateSecret('test')
-    const { hash, contractTransaction } = await rskRegistrar.commitToRegister('luca.rsk', await rnsOwner.getAddress(), salt)
+    const { hash, contractTransaction } = await rskRegistrar.commitToRegister('lucachaco', await rnsOwner.getAddress(), salt)
     console.log({ hash })
     console.log({ salt })
     const commitToRegisterReceipt = await contractTransaction.wait()
     console.log({ commitToRegisterReceipt })
     /*    const available = await rskRegistrar.available('new-domain.rsk')
     console.log({ available }) */
-    /* const canRevealResponse = await rskRegistrar.canReveal(hash)
-    console.log({ canRevealResponse }) */
-    /* const registerTx = await rskRegistrar.register('luca.rsk', await rnsOwner.getAddress(), salt)
-    const registerReceipt = registerTx.wait() */
-    /* console.log({ registerReceipt }) */
+    await advanceTime()
+    const canRevealResponse = await rskRegistrar.canReveal(hash)
+    expect(canRevealResponse).toEqual(true)
+    console.log({ canRevealResponse })
+    const registerTx = await rskRegistrar.register('lucachaco', await rnsOwner.getAddress(), salt)
+    const registerReceipt = registerTx.wait()
+    console.log({ registerReceipt })
   })
 })
+
+const advanceTime = async () => {
+  const rpcUrl = 'http://localhost:8545'
+  const provider = new providers.JsonRpcProvider(rpcUrl)
+  const evmIncreaseTime = await provider.send('evm_increaseTime', [1000])
+  console.log(evmIncreaseTime)
+  const evmMine = await provider.send('evm_mine', [])
+  console.log(evmMine)
+}
